+++ src/topBar/index.tsx
@@ -1,18 +1,17 @@
 import React, { Fragment } from 'react'
 import cc from 'classcat'
 
-import prefix from '_utils'
 import style from 'topBar/style'
 
 interface TopBarProps {
-  readonly className?: Classcat.Class,
-  readonly leftItem?: JSX.Element,
-  readonly rightItem?: JSX.Element,
-  readonly centerItem?: JSX.Element,
-  readonly fixed?: boolean,
-  readonly bgTransparent?: boolean,
-  readonly bgShadedTransparent?: boolean,
-  readonly innerWrapperClassName?: Classcat.Class,
+  readonly className?: Classcat.Class
+  readonly leftItem?: JSX.Element
+  readonly rightItem?: JSX.Element
+  readonly centerItem?: JSX.Element
+  readonly fixed?: boolean
+  readonly bgTransparent?: boolean
+  readonly bgShadedTransparent?: boolean
+  readonly innerWrapperClassName?: Classcat.Class
 }
 
 const TopBar = ({
@@ -24,7 +23,7 @@ const TopBar = ({
   rightItem,
   centerItem,
   innerWrapperClassName = null,
-}:TopBarProps) => {
+}: TopBarProps) => {
   const Wrapper = Boolean(innerWrapperClassName) ? 'div' : Fragment
   const wrapperProps: { className?: string } = {}
   if (Boolean(innerWrapperClassName)) {
@@ -43,27 +42,9 @@ const TopBar = ({
       ])}
     >
       <Wrapper {...wrapperProps}>
-        {
-          leftItem ? (
-            <div className={cc({ 'kirk-topBar-left': true })}>
-              {leftItem}
-            </div>
-          ) : null
-        }
-        {
-          centerItem ? (
-            <div className={cc({ 'kirk-topBar-center': true })}>
-              {centerItem}
-            </div>
-          ) : null
-        }
-        {
-          rightItem ? (
-            <div className={cc({ 'kirk-topBar-right': true })}>
-              {rightItem}
-            </div>
-          ) : null
-        }
+        {leftItem ? <div className={cc({ 'kirk-topBar-left': true })}>{leftItem}</div> : null}
+        {centerItem ? <div className={cc({ 'kirk-topBar-center': true })}>{centerItem}</div> : null}
+        {rightItem ? <div className={cc({ 'kirk-topBar-right': true })}>{rightItem}</div> : null}
       </Wrapper>
       <style jsx>{style}</style>
     </header>